pyproject.toml
[project]
name = "libibt"
version = "0.1.0"
description = "Library for reading iRacing IBT telemetry files"
license = "MIT"
authors = [
    {name = "Christopher Dewan", email = "chris.dewan@m3rlin.net"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Rust",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.26.0; python_version < '3.13'",
    "numpy>=2.1.0; python_version >= '3.13' and python_version < '3.14'",
    "numpy>=2.4.0; python_version >= '3.14'",
    "pyarrow>=18.1.0; python_version < '3.14'",
    "pyarrow>=22.0.0; python_version >= '3.14'",
]

[project.urls]
Homepage = "https://github.com/m3rlin45/libibt"
Repository = "https://github.com/m3rlin45/libibt"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=25.11.0,<26.0.0",
    "mypy>=1.18.2",
    "construct>=2.10.0",
    "pyyaml>=6.0",
    "pytest-xdist>=3.8.0",
    "maturin>=1.11.5",
    "pyirsdk>=1.3.5",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.mypy_cache
  | \.venv
  | build
  | dist
  | reference
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
exclude = [
    "^build/",
    "^dist/",
    "^reference/",
]

[[tool.mypy.overrides]]
module = "numpy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyarrow.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "construct.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.maturin]
python-source = "src"
module-name = "libibt._libibt_rs"
features = ["pyo3/extension-module"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["maturin>=1.11.5"]
build-backend = "maturin"
